requirements.txt
pymupdf
pdfminer.six
tqdm
pydantic>=2
rank-bm25
rapidfuzz
sentence-transformers
faiss-cpu
numpy
transformers
